arena component is divided into separate components
