+++ src/UI/Footer/Footer.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import styles from './Footer.module.sass'
+
+const Footer = () => {
+  return (
+    <div className={styles.footer}/>
+  )
+}
+
+export default Footer
