+++ src/UI/Footer/Footer.module.sass
@@ -1,0 +1,4 @@
+.footer
+  width: 100%
+  height: 50px
+  background-color: darkslategrey
