+++ src/UI/Loader/Loader.module.sass
@@ -1,0 +1,38 @@
+.lds_ring
+  display: inline-block
+  position: relative
+
+.lds_ring div
+  box-sizing: border-box
+  display: block
+  position: absolute
+  width: 80%
+  height: 80%
+  margin: 50%
+  border-style: solid
+  border-radius: 50%
+  animation: lds_ring 1.2s cubic-bezier(0.5, 0, 0.5, 1) infinite
+  border-color: black transparent transparent transparent
+
+.lds_ring div:nth-child(1)
+  animation-delay: -0.45s
+
+.lds_ring div:nth-child(2)
+  animation-delay: -0.3s
+
+.lds_ring div:nth-child(3)
+  animation-delay: -0.15s
+
+.center
+  display: flex
+  align-items: center
+  justify-content: center
+  width: 100%
+  height: 100%
+
+@keyframes lds_ring
+  0%
+    transform: rotate(0deg)
+
+  100%
+    transform: rotate(360deg)
