+++ src/components/Container/Partition/Arena/Arena.module.sass
@@ -5,349 +5,3 @@
   width: 100%
   padding: 0 calc((100% - 980px)/2) 20px
   background-color: whitesmoke
-
-.arena_content
-  display: flex
-  justify-content: space-between
-  width: 100%
-  padding: 10px 0
-
-.images
-  height: 400px
-  width: 400px
-
-.information
-  width: 65%
-  height: 100%
-  font-size: 19px
-
-.additional
-  display: flex
-  flex-direction: column
-  margin: 30px 0
-  width: 70%
-  &_item
-    width: 100%
-    position: relative
-    margin-bottom: 10px
-    padding-bottom: 10px
-    word-break: break-all
-    hyphens: auto
-    &:after
-      position: absolute
-      content: ''
-      background-color: grey
-      height: 1px
-      bottom: 0
-      left: 0
-      z-index: 1000
-      width: 100%
-
-.address, .working_hours
-  margin-top: 10px
-  white-space: nowrap
-  overflow: hidden
-  -ms-text-overflow: ellipsis
-  text-overflow: ellipsis
-  &:hover
-    overflow: visible
-
-.social
-  display: flex
-  margin-top: 15px
-
-.instagram, .telegram, .vk, .youtube
-  display: flex
-  align-items: center
-  justify-content: center
-  width: 32px
-  height: 32px
-  margin-right: 20px
-  background-color: grey
-  -webkit-border-radius: 50%
-  -moz-border-radius: 50%
-  border-radius: 50%
-
-.site
-  margin-top: 10px
-  white-space: nowrap
-  overflow: hidden
-  -ms-text-overflow: ellipsis
-  text-overflow: ellipsis
-  &:hover
-    overflow: visible
-
-.site_link
-  text-decoration: none
-  color: blue
-  &:hover
-    text-decoration: underline
-
-@keyframes marquee
-  from
-    text-indent: 100%
-  to
-    text-indent: 0
-
-.map
-  height: 400px
-  -webkit-border-radius: 5px
-  -moz-border-radius: 5px
-  border-radius: 5px
-  margin: 10px 0
-  overflow: hidden
-  width: 100%
-
-.rental
-  width: 100%
-
-.title
-  text-decoration: underline
-  margin: 15px 0
-
-.rented
-  display: flex
-  &_hours
-    display: flex
-    justify-content: space-between
-    align-content: space-between
-    flex-wrap: wrap
-    width: 400px
-    height: 200px
-  &_info
-    display: flex
-    flex-direction: column
-    justify-content: space-between
-    position: relative
-    flex-grow: 1
-  &_button
-    width: 110px
-    height: 40px
-    margin-top: 20px
-    border: none
-    background-color: #44944a
-    opacity: .8
-    transition: .3s all
-    cursor: pointer
-    &:hover
-      opacity: 1
-  &_error
-    position: absolute
-    bottom: 10px
-    right: 10px
-    color: #e32636
-    font-weight: bold
-
-.booked
-  &_date
-    display: block
-    margin-top: 30px
-    font-size: 19px
-  &_hours
-    display: block
-    font-size: 19px
-.datepicker
-  margin: 10px 0
-
-.hour
-  position: absolute
-  width: 0
-  height: 0
-  opacity: 0
-  &:checked
-    ~ .hour_span
-      background-color: #E32636
-    ~ .hour_span:after
-      display: block
-  &_label
-    width: 40px
-    height: 40px
-    z-index: 2
-    opacity: .85
-  &_span
-    display: flex
-    align-items: center
-    justify-content: center
-    position: absolute
-    top: 0
-    left: 0
-    width: 40px
-    height: 40px
-    background-color: #44944A
-    transition: .3s all
-    z-index: 2
-    opacity: .8
-    &:after
-      content: ''
-      position: absolute
-      display: none
-      width: 8px
-      height: 20px
-      border: solid white
-      border-width: 0 6px 6px 0
-      transform: rotate(45deg)
-      margin-bottom: 5px
-  &[disabled]
-    ~ .hour_interval
-      color: black
-      opacity: .2 !important
-    ~ .hour_span
-      opacity: .2
-  &_container
-    display: flex
-    flex-direction: column
-    justify-content: center
-    align-items: center
-    position: relative
-    cursor: pointer
-    margin-right: calc(100%/6 - 40px)
-    &:hover
-      .hour_span
-        background-color: darken(#44944A, 10%)
-      .hour_interval
-        opacity: unset
-  &_interval
-    display: flex
-    flex-direction: column
-    text-align: center
-    position: absolute
-    font-size: 12px
-    opacity: .5
-    user-select: none
-  &:nth-child(6n)
-    margin-right: 0
-
-.checkboxes
-  display: flex
-  width: 100%
-  justify-content: space-between
-
-.checkbox
-  &_rented, &_unbooked, &_booked
-    display: flex
-    align-items: center
-    justify-content: center
-    position: relative
-    width: 40px
-    height: 40px
-    margin-right: 10px
-  &_rented, &_booked
-    background-color: #E32636
-    opacity: .2
-    &:after
-      content: ''
-      position: absolute
-      width: 8px
-      height: 20px
-      border: solid white
-      border-width: 0 6px 6px 0
-      transform: rotate(45deg)
-      margin-bottom: 5px
-  &_unbooked
-    background-color: #44944A
-    opacity: .8
-  &_booked
-    opacity: .8
-  &_info
-    display: flex
-    align-items: center
-
-
-
-.lds_ring, .lds_ring_small
-  display: inline-block
-  position: relative
-  width: 80px
-  height: 80px
-
-.lds_ring_small
-  width: 20px
-  height: 20px
-
-.lds_ring div, .lds_ring_small div
-  box-sizing: border-box
-  display: block
-  position: absolute
-  width: 64px
-  height: 64px
-  margin: 8px
-  border: 8px solid
-  border-radius: 50%
-  animation: lds_ring 1.2s cubic-bezier(0.5, 0, 0.5, 1) infinite
-  border-color: black transparent transparent transparent
-
-.lds_ring_small div
-  width: 16px
-  height: 16px
-  margin: 2px
-  border: 2px solid
-.lds_ring div:nth-child(1), .lds_ring_small div:nth-child(1)
-  animation-delay: -0.45s
-
-.lds_ring div:nth-child(2), .lds_ring_small div:nth-child(2)
-  animation-delay: -0.3s
-
-.lds_ring div:nth-child(3), .lds_ring_small div:nth-child(3)
-  animation-delay: -0.15s
-
-.center
-  text-align: center
-
-@keyframes lds_ring
-  0%
-    transform: rotate(0deg)
-
-  100%
-    transform: rotate(360deg)
-
-
-
-
-.commands
-  width: 100%
-  margin-top: 40px
-  &_list
-    padding: 15px 5px
-    border: 1px solid #000
-    max-height: 700px
-    overflow-y: auto
-
-
-.command
-  &_card
-    display: flex
-    justify-content: flex-start
-    border: 1px solid black
-    height: 130px
-    padding: 5px
-    margin-bottom: 10px
-    &:nth-child(2n)
-      background-color: #c1c1c1
-    &:last-child
-      margin-bottom: 0
-  &_image
-    object-fit: cover
-    &_container
-      height: 100%
-      width: 120px
-      margin-right: 10px
-  &_info
-    display: flex
-    flex-direction: column
-    justify-content: space-between
-    flex-grow: 1
-  &_name
-    font-size: 22px
-    margin-bottom: 5px
-    text-decoration: none
-  &_phone, &_email
-    display: inline
-    color: #000
-    text-decoration: none
-    //&:hover
-    //  text-decoration: underline
-
-.footer
-  width: 100%
-  height: 50px
-  background-color: darkslategrey
