+++ src/components/Container/Partition/Arena/ArenaMap/ArenaMap.module.sass
@@ -1,0 +1,8 @@
+.map
+  height: 400px
+  -webkit-border-radius: 5px
+  -moz-border-radius: 5px
+  border-radius: 5px
+  margin: 10px 0
+  overflow: hidden
+  width: 100%
