+++ src/components/Container/Partition/Arena/Commands/Commands.module.sass
@@ -1,0 +1,62 @@
+.commands
+  width: 100%
+  margin-top: 40px
+  &_list
+    padding: 15px 5px
+    border: 1px solid #000
+    max-height: 700px
+    overflow-y: auto
+
+.title
+  text-decoration: underline
+  margin: 15px 0
+
+.command
+  &_card
+    display: flex
+    justify-content: flex-start
+    border: 1px solid black
+    height: 130px
+    padding: 5px
+    margin-bottom: 10px
+    &:nth-child(2n)
+      background-color: #c1c1c1
+    &:last-child
+      margin-bottom: 0
+  &_image
+    width: 100%
+    height: 100%
+    object-fit: cover
+    &_container
+      height: 100%
+      width: 120px
+      margin-right: 10px
+  &_info
+    display: flex
+    flex-direction: column
+    justify-content: space-between
+    flex-grow: 1
+  &_name
+    font-size: 22px
+    margin-bottom: 5px
+    text-decoration: none
+  &_phone, &_email
+    display: inline
+    color: #000
+    text-decoration: none
+
+.social
+  display: flex
+  margin-top: 15px
+
+.instagram, .telegram, .vk, .youtube
+  display: flex
+  align-items: center
+  justify-content: center
+  width: 30px
+  height: 30px
+  margin-right: 15px
+  background-color: grey
+  -webkit-border-radius: 50%
+  -moz-border-radius: 50%
+  border-radius: 50%
