+++ src/components/Container/Partition/Arena/Rental/Rental.module.sass
@@ -1,0 +1,12 @@
+.rental
+  width: 100%
+
+.title
+  text-decoration: underline
+  margin: 15px 0
+
+.rented
+  display: flex
+
+.datepicker
+  margin: 10px 0
