+++ src/components/Container/Partition/Arena/Rental/RentedHours/RentedHours.module.sass
@@ -1,0 +1,114 @@
+.rented
+  display: flex
+  &_hours
+    display: flex
+    justify-content: space-between
+    align-content: space-between
+    flex-wrap: wrap
+    width: 400px
+    height: 200px
+
+.hour
+  position: absolute
+  width: 0
+  height: 0
+  opacity: 0
+  &:checked
+    ~ .hour_span
+      background-color: #E32636
+    ~ .hour_span:after
+      display: block
+  &_label
+    width: 40px
+    height: 40px
+    z-index: 2
+    opacity: .85
+  &_span
+    display: flex
+    align-items: center
+    justify-content: center
+    position: absolute
+    top: 0
+    left: 0
+    width: 40px
+    height: 40px
+    background-color: #44944A
+    transition: .3s all
+    z-index: 2
+    opacity: .8
+    &:after
+      content: ''
+      position: absolute
+      display: none
+      width: 8px
+      height: 20px
+      border: solid white
+      border-width: 0 6px 6px 0
+      transform: rotate(45deg)
+      margin-bottom: 5px
+  &[disabled]
+    ~ .hour_interval
+      color: black
+      opacity: .2 !important
+    ~ .hour_span
+      opacity: .2
+  &_container
+    display: flex
+    flex-direction: column
+    justify-content: center
+    align-items: center
+    position: relative
+    cursor: pointer
+    margin-right: calc(100%/6 - 40px)
+    &:hover
+      .hour_span
+        background-color: darken(#44944A, 10%)
+      .hour_interval
+        opacity: unset
+  &_interval
+    display: flex
+    flex-direction: column
+    text-align: center
+    position: absolute
+    font-size: 12px
+    opacity: .5
+    user-select: none
+  &:nth-child(6n)
+    margin-right: 0
+
+.lds_ring
+  display: inline-block
+  position: relative
+  width: 80px
+  height: 80px
+
+.lds_ring div
+  box-sizing: border-box
+  display: block
+  position: absolute
+  width: 64px
+  height: 64px
+  margin: 8px
+  border: 8px solid
+  border-radius: 50%
+  animation: lds_ring 1.2s cubic-bezier(0.5, 0, 0.5, 1) infinite
+  border-color: black transparent transparent transparent
+
+.lds_ring div:nth-child(1)
+  animation-delay: -0.45s
+
+.lds_ring div:nth-child(2)
+  animation-delay: -0.3s
+
+.lds_ring div:nth-child(3)
+  animation-delay: -0.15s
+
+.center
+  text-align: center
+
+@keyframes lds_ring
+  0%
+    transform: rotate(0deg)
+
+  100%
+    transform: rotate(360deg)
